Move out NearQueryNode, ONearQueryNode, PhraseQueryNode and SameElementQueryNode to separate files.
