+++ searchlib/src/tests/query/streaming/CMakeLists.txt
@@ -17,3 +17,21 @@ vespa_add_executable(searchlib_query_streaming_hit_iterator_pack_test_app TEST
     GTest::gtest
 )
 vespa_add_test(NAME searchlib_query_streaming_hit_iterator_pack_test_app COMMAND searchlib_query_streaming_hit_iterator_pack_test_app)
+
+vespa_add_executable(searchlib_query_streaming_same_element_query_node_test_app TEST
+    SOURCES
+    same_element_query_node_test.cpp
+    DEPENDS
+    searchlib
+    GTest::gtest
+)
+vespa_add_test(NAME searchlib_query_streaming_same_element_query_node_test_app COMMAND searchlib_query_streaming_same_element_query_node_test_app)
+
+vespa_add_executable(searchlib_query_streaming_phrase_query_node_test_app TEST
+    SOURCES
+    phrase_query_node_test.cpp
+    DEPENDS
+    searchlib
+    GTest::gtest
+)
+vespa_add_test(NAME searchlib_query_streaming_phrase_query_node_test_app COMMAND searchlib_query_streaming_phrase_query_node_test_app)
