+++ searchlib/src/vespa/searchlib/query/streaming/near_query_node.h
@@ -1,0 +1,26 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#pragma once
+
+#include "query.h"
+
+namespace search::streaming {
+
+/**
+   N-ary Near operator. All terms must be within the given distance.
+*/
+class NearQueryNode : public AndQueryNode
+{
+public:
+    NearQueryNode() noexcept : AndQueryNode("NEAR"), _distance(0) { }
+    explicit NearQueryNode(const char * opName) noexcept : AndQueryNode(opName), _distance(0) { }
+    bool evaluate() const override;
+    void distance(size_t dist)       { _distance = dist; }
+    size_t distance()          const { return _distance; }
+    void visitMembers(vespalib::ObjectVisitor &visitor) const override;
+    bool isFlattenable(ParseItem::ItemType type) const override { return type == ParseItem::ITEM_NOT; }
+private:
+    size_t _distance;
+};
+
+}
