+++ searchlib/src/vespa/searchlib/query/streaming/query.h
@@ -108,69 +108,6 @@ public:
 };
 
 /**
-   N-ary phrase operator. All terms must be satisfied and have the correct order
-   with distance to next term equal to 1.
-*/
-class PhraseQueryNode : public AndQueryNode
-{
-public:
-    PhraseQueryNode() noexcept : AndQueryNode("PHRASE"), _fieldInfo(32) { }
-    bool evaluate() const override;
-    const HitList & evaluateHits(HitList & hl) const override;
-    void getPhrases(QueryNodeRefList & tl) override;
-    void getPhrases(ConstQueryNodeRefList & tl) const override;
-    const QueryTerm::FieldInfo & getFieldInfo(size_t fid) const { return _fieldInfo[fid]; }
-    size_t getFieldInfoSize() const { return _fieldInfo.size(); }
-    bool isFlattenable(ParseItem::ItemType type) const override { return type == ParseItem::ITEM_NOT; }
-    void addChild(QueryNode::UP child) override;
-private:
-    mutable std::vector<QueryTerm::FieldInfo> _fieldInfo;
-    void updateFieldInfo(size_t fid, size_t offset, size_t fieldLength) const;
-#if WE_EVER_NEED_TO_CACHE_THIS_WE_MIGHT_WANT_SOME_CODE_HERE
-    HitList _cachedHitList;
-    bool    _evaluated;
-#endif
-};
-
-class SameElementQueryNode : public AndQueryNode
-{
-public:
-    SameElementQueryNode() noexcept : AndQueryNode("SAME_ELEMENT") { }
-    bool evaluate() const override;
-    const HitList & evaluateHits(HitList & hl) const override;
-    bool isFlattenable(ParseItem::ItemType type) const override { return type == ParseItem::ITEM_NOT; }
-    void addChild(QueryNode::UP child) override;
-};
-
-/**
-   N-ary Near operator. All terms must be within the given distance.
-*/
-class NearQueryNode : public AndQueryNode
-{
-public:
-    NearQueryNode() noexcept : AndQueryNode("NEAR"), _distance(0) { }
-    explicit NearQueryNode(const char * opName) noexcept : AndQueryNode(opName), _distance(0) { }
-    bool evaluate() const override;
-    void distance(size_t dist)       { _distance = dist; }
-    size_t distance()          const { return _distance; }
-    void visitMembers(vespalib::ObjectVisitor &visitor) const override;
-    bool isFlattenable(ParseItem::ItemType type) const override { return type == ParseItem::ITEM_NOT; }
-private:
-    size_t _distance;
-};
-
-/**
-   N-ary Ordered near operator. The terms must be in order and the distance between
-   the first and last must not exceed the given distance.
-*/
-class ONearQueryNode : public NearQueryNode
-{
-public:
-    ONearQueryNode() noexcept : NearQueryNode("ONEAR") { }
-    bool evaluate() const override;
-};
-
-/**
    Query packages the query tree. The usage pattern is like this.
    Construct the tree with the correct tree description.
    Get the leaf nodes and populate them with the term occurences.
